Move assets stuff into its own package
