+++ flows/definition/testdata/assets.json
@@ -1,0 +1,186 @@
+[
+    {
+        "type": "flow",
+        "url": "http://testserver/assets/flow/76f0a02f-3b75-4b86-9064-e9195e1b3a02",
+        "content": {
+            "uuid": "76f0a02f-3b75-4b86-9064-e9195e1b3a02",
+            "name": "Test Flow",
+            "language": "eng",
+            "nodes": [
+                {
+                    "uuid": "a58be63b-907d-4a1a-856b-0bb5579d7507",
+                    "actions": [
+                        {
+                            "uuid": "9487a60e-a6ef-4a88-b35d-894bfe074144",
+                            "type": "start_flow",
+                            "flow": {
+                                "uuid": "b7cf0d83-f1c9-411c-96fd-c511a4cfa86d",
+                                "name": "Collect Language"
+                            }
+                        },
+                        {
+                            "uuid": "ad154980-7bf7-4ab8-8728-545fd6378912",
+                            "type": "add_input_labels",
+                            "labels": [
+                                {
+                                    "uuid": "3f65d88a-95dc-4140-9451-943e94e06fea",
+                                    "name": "Spam"
+                                }
+                            ]
+                        },
+                        {
+                            "uuid": "ad154980-7bf7-4ab8-8728-545fd6378912",
+                            "type": "add_contact_groups",
+                            "groups": [
+                                {
+                                    "uuid": "2aad21f6-30b7-42c5-bd7f-1b720c154817",
+                                    "name": "Survey Audience"
+                                }
+                            ]
+                        },
+                        {
+                            "uuid": "7bd8b3bf-0a3c-4928-bc46-df416e77ddf4",
+                            "type": "set_contact_field",
+                            "field": {
+                                "key": "first_name",
+                                "label": "First Name"
+                            },
+                            "value": "Bob"
+                        },
+                        {
+                            "uuid": "3248a064-bc42-4dff-aa0f-93d85de2f600",
+                            "type": "set_contact_channel",
+                            "channel": {
+                                "uuid": "57f1078f-88aa-46f4-a59a-948a5739c03d",
+                                "name": "Twilio Channel"
+                            }
+                        }, 
+                        {
+                            "uuid": "3248a064-bc42-4dff-aa0f-93d85de2f600",
+                            "type": "set_run_result",
+                            "name": "Favorite Color",
+                            "value": "red",
+                            "category": "Red"
+                        }
+                    ],
+                    "wait": {
+                        "type": "msg",
+                        "timeout": 600
+                    },
+                    "exits": [
+                        {
+                            "uuid": "2d481ce6-efcf-4898-a825-f76208e32f2a",
+                            "destination_node_uuid": "3bfdb92a-3ffd-48e5-adbd-2ad8f8869437"
+                        }
+                    ]
+                },
+                {
+                    "uuid": "3bfdb92a-3ffd-48e5-adbd-2ad8f8869437",
+                    "actions": [],
+                    "wait": {
+                        "type": "msg",
+                        "timeout": 600
+                    } 
+                }
+            ]
+        }
+    },
+    {
+        "type": "flow",
+        "url": "http://testserver/assets/flow/b7cf0d83-f1c9-411c-96fd-c511a4cfa86d",
+        "content": {
+            "uuid": "b7cf0d83-f1c9-411c-96fd-c511a4cfa86d",
+            "name": "Collect Language",
+            "nodes": [{
+                "uuid": "d9dba561-b5ee-4f62-ba44-60c4dc242b84",
+                "actions": []
+            }]
+        }
+    },
+    {
+        "type": "field",
+        "url": "http://testserver/assets/field",
+        "content": [
+            {
+                "key": "first_name",
+                "label": "First Name",
+                "value_type": "text"
+            },
+            {
+                "key": "activation_token",
+                "label": "Activation Token",
+                "value_type": "text"
+            },
+            {
+                "key": "gender",
+                "label": "Gender",
+                "value_type": "text"
+            },
+            {
+                "key": "age",
+                "label": "Age",
+                "value_type": "decimal"
+            },
+            {
+                "key": "state",
+                "label": "State",
+                "value_type": "text"
+            },
+            {
+                "key": "joined",
+                "label": "Joined On",
+                "value_type": "datetime"
+            }
+        ],
+        "is_set": true
+    },
+    {
+        "type": "group",
+        "url": "http://testserver/assets/group",
+        "content": [
+            {
+                "uuid": "2aad21f6-30b7-42c5-bd7f-1b720c154817",
+                "name": "Survey Audience"
+            },
+            {
+                "uuid": "d7ff4872-9238-452f-9d38-2f558fea89e0",
+                "name": "Azuay State"
+            },
+            {
+                "uuid": "435a8fad-6780-49ab-93ec-443c20386df8",
+                "name": "Males",
+                "query": "gender = Male"
+            }
+        ],
+        "is_set": true
+    },
+    {
+        "type": "channel",
+        "url": "http://testserver/assets/channel",
+        "content": [
+            {
+                "uuid": "57f1078f-88aa-46f4-a59a-948a5739c03d",
+                "name": "Nexmo",
+                "address": "+12345671111",
+                "schemes": ["tel"],
+                "roles": ["send", "receive"]
+            }
+        ],
+        "is_set": true
+    },
+    {
+        "type": "label",
+        "url": "http://testserver/assets/label",
+        "content": [
+            {
+                "uuid": "3f65d88a-95dc-4140-9451-943e94e06fea",
+                "name": "Spam"
+            },
+            {
+                "uuid": "b017c07a-d35b-4da4-8917-3bf8bff80168",
+                "name": "Azuay Messages"
+            }
+        ],
+        "is_set": true
+    }
+]
